+++ Tests/CommandTest.php
@@ -19,6 +19,12 @@ class CommandTest extends TestCase
 
         $this->artisan('dynamic:migrate one --path=database/migrations')->assertSuccessful();
 
+        $this->artisan('dynamic:migrate one --seed --path=database/migrations')->assertSuccessful();
+
+        $this->artisan('dynamic:migrate one --refresh --path=database/migrations')->assertSuccessful();
+
+        $this->artisan('dynamic:migrate one --fresh --path=database/migrations')->assertSuccessful();
+
         $this->artisan('vendor:publish --tag=ddc-config')->assertSuccessful();
 
         $this->artisan('vendor:publish --tag=ddc-migrations')->assertSuccessful();