+++ src/Console/Commands/EnvDatabaseConfigMigrateCommand.php
@@ -19,9 +19,11 @@ class EnvDatabaseConfigMigrateCommand extends Command
                             {name : The name of the new database connection}
                             {postfix : The postfix for the database configuration}
                             {--P|--path= : The path where the database migration files are kept}
+                            {--seed : Running seeders}
                             {--seeder= : Running a single seeder class}
                             {--refresh : Refreshing all migration}
-                            {--rollback : Reverting migrations}';
+                            {--rollback : Reverting migrations}
+                            {--fresh : Re-run all migrations afresh}';
 
     /**
      * The console command description.
@@ -54,14 +56,21 @@ class EnvDatabaseConfigMigrateCommand extends Command
 
     private function execMigrationCommands(string $name, null|string $path = '')
     {
+        if ($this->option('rollback')) {
+            $this->call('migrate:rollback', ['--database' => $name, '--path' => $path]);
+            return;
+        }
+
         if ($this->option('refresh')) {
             $this->call('migrate:refresh', ['--database' => $name, '--path' => $path]);
+        } elseif ($this->option('fresh')) {
+            $this->call('migrate:fresh', ['--database' => $name, '--path' => $path]);
         } else {
             $this->call('migrate', ['--database' => $name, '--path' => $path]);
         }
 
-        if ($this->option('rollback')) {
-            $this->call('migrate:rollback', ['--database' => $name, '--path' => $path]);
+        if ($this->option('seed')) {
+            $this->call('db:seed');
             return;
         }